[refactor] 출석 API 리팩토링
what : 출석 API의 URI 주소를 수정하였습니다. 추가적으로 예외가 잘못 던져지는 부분들을 수정하였습니다.
