+++ Backend/src/main/java/com/github/backend/service/attendance/impl/AttendanceServiceImpl.java
@@ -2,14 +2,15 @@ package com.github.backend.service.attendance.impl;
 
 import static com.github.backend.dto.attendance.AttendanceDto.Info;
 import static com.github.backend.exception.attendance.code.AttendanceErrorCode.ALREADY_CHECK_ATTENDANCE;
-import static com.github.backend.exception.attendance.code.AttendanceErrorCode.MEMBER_NOT_EXISTS;
+import static com.github.backend.exception.member.code.MemberErrorCode.MEMBER_NOT_EXISTS;
 import static java.util.stream.Collectors.toList;
 
-import com.github.backend.exception.attendance.AttendanceException;
 import com.github.backend.dto.attendance.AttendanceDto.GetRequest;
+import com.github.backend.exception.attendance.AttendanceException;
+import com.github.backend.exception.member.MemberException;
 import com.github.backend.persist.attendance.Attendance;
-import com.github.backend.persist.member.Member;
 import com.github.backend.persist.attendance.repository.AttendanceRepository;
+import com.github.backend.persist.member.Member;
 import com.github.backend.persist.member.repository.MemberRepository;
 import com.github.backend.service.attendance.AttendanceService;
 import java.time.LocalDate;
@@ -31,7 +32,7 @@ public class AttendanceServiceImpl implements AttendanceService {
 	@Override
 	public void checkTodayAttendance(String email) {
 		Member member = memberRepository.findByEmail(email)
-			.orElseThrow(() -> new AttendanceException(MEMBER_NOT_EXISTS));
+			.orElseThrow(() -> new MemberException(MEMBER_NOT_EXISTS));
 
 		LocalDate today = LocalDate.now();
