+++ Backend/src/main/java/com/github/backend/web/attendance/AttendanceController.java
@@ -1,4 +1,4 @@
-package com.github.backend.web.member.controller;
+package com.github.backend.web.attendance;
 
 import com.github.backend.dto.common.Result;
 import com.github.backend.dto.attendance.AttendanceDto;
@@ -19,13 +19,13 @@ public class AttendanceController {
 
 	private final AttendanceService attendanceService;
 
-	@PostMapping("/member/attendance")
+	@PostMapping("/members/me/attendances/daily")
 	public ResponseEntity<Result<?>> checkDailyAttendance(@AuthenticationPrincipal MemberInfo memberInfo) {
 		attendanceService.checkTodayAttendance(memberInfo.getEmail());
 		return ResponseEntity.ok().body(new Result<>(200,true,""));
 	}
 
-	@GetMapping("/member/attendance")
+	@GetMapping("/members/me/attendances")
 	public ResponseEntity<Result<?>> getAttendances(@AuthenticationPrincipal MemberInfo memberInfo,
 		AttendanceDto.GetRequest request) {
