+++ Backend/src/test/java/com/github/backend/service/attendance/impl/AttendanceServiceImplTest.java
@@ -1,4 +1,4 @@
-package com.github.backend.service.impl;
+package com.github.backend.service.attendance.impl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,17 +8,16 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
 
-import com.github.backend.exception.attendance.AttendanceException;
 import com.github.backend.dto.attendance.AttendanceDto;
 import com.github.backend.dto.attendance.AttendanceDto.GetRequest;
+import com.github.backend.exception.attendance.AttendanceException;
+import com.github.backend.exception.attendance.code.AttendanceErrorCode;
 import com.github.backend.exception.member.MemberException;
 import com.github.backend.exception.member.code.MemberErrorCode;
 import com.github.backend.persist.attendance.Attendance;
-import com.github.backend.persist.member.Member;
 import com.github.backend.persist.attendance.repository.AttendanceRepository;
+import com.github.backend.persist.member.Member;
 import com.github.backend.persist.member.repository.MemberRepository;
-import com.github.backend.exception.attendance.code.AttendanceErrorCode;
-import com.github.backend.service.attendance.impl.AttendanceServiceImpl;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
