+++ Backend/src/test/java/com/github/backend/web/AdminControllerTest.java
@@ -17,12 +17,13 @@ import com.github.backend.dto.admin.CreateAdmin;
 import com.github.backend.dto.admin.RegisterAdminOutputDto;
 import com.github.backend.dto.course.CourseInfoOutputDto;
 import com.github.backend.dto.course.RegisterCourse;
+import com.github.backend.dto.course.UpdateCourse;
 import com.github.backend.dto.gamer.GamerInfoOutputDto;
 import com.github.backend.dto.gamer.RegisterGamer;
-import com.github.backend.dto.course.UpdateCourse;
 import com.github.backend.dto.gamer.RegisterGamerOutputDto;
 import com.github.backend.dto.gamer.UpdateGamer;
-
+import com.github.backend.persist.admin.Admin;
+import com.github.backend.persist.course.Course;
 import com.github.backend.persist.gamer.Gamer;
 import com.github.backend.persist.member.type.Role;
 import com.github.backend.security.jwt.JwtAuthenticationProvider;
